requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
markdownify==0.13.1
mcp==1.9.4
mdurl==0.1.2
numpy==1.26.3
pandas==2.3.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jobspy==1.1.80
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.1
tls-client==1.0.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
